Predictions from GP, DGP, or linked (D)GP emulators

This function implements prediction from GP, DGP, or linked (D)GP emulators.

Usage

predict(obj, x, method = "mean_var", full_layer = FALSE, sample_size = 50)

Arguments

obj

a GP, DGP, or linked (D)GP object produced by gp() (after applying train()), emulator(), or lgp().

x

a matrix where each row is an input testing data point and each column is an input dimension.

method

the prediction approach: mean-variance ("mean_var") or sampling ("sampling") approach. Defaults to "mean_var".

full_layer

whether to output the predictions of all layers. Defaults to FALSE. Only used when obj is for DGP and linked (D)GP.

sample_size

the number of samples to draw for each given imputation if method = "sampling". Defaults to 50.

Value

(GP case) If obj is produced by train() on a object built with gp():

if method = "mean_var": a named list is returned that contains two matrices named mean for the predictive means and var for the predictive variances. Each matrix has only one column with its rows corresponding to testing positions (i.e., row of x).

if method = "sampling": a matrix is returned whose rows corresponding to testing positions and columns corresponding to sample_size number of samples drawn from the predictive distribution of GP.

(DGP case) If obj is produced by emulator():

if method = "mean_var" and full_layer = FALSE: a named list is returned. The list contains two matrices named mean for the predictive means and var for the predictive variances respectively. Each matrix has its rows corresponding to testing positions and columns corresponding to DGP global output dimensions (i.e., the number of GP/likelihood nodes in the final layer).

if method = "mean_var" and full_layer = TRUE: a named list is returned. The list contains two sub-lists named mean for the predictive means and var for the predictive variances respectively. Each sub-list contains L (i.e., the number of layers) matrices named layer1, layer2,..., layerL. Each matrix has its rows corresponding to testing positions and columns corresponding to output dimensions (i.e., the number of GP nodes from the associated layer and in case of the final layer, it may be the number of the likelihood nodes).

if method = "sampling" and full_layer = FALSE: a named list is returned. The list contains D (i.e., the number of GP/likelihood nodes in the final layer) matrices named output1, output2,..., outputD. Each matrix has its rows corresponding to testing positions and columns corresponding to samples of size: N * sample_size, where N is the number of imputations specified in emulator().

if method = "sampling" and full_layer = TRUE: a named list is returned. The list contains L (i.e., the number of layers) sub-lists named layer1, layer2,..., layerL. Each sub-list represents samples drawn from the GPs/likelihoods in the corresponding layers, and contains D (i.e., the number of GP nodes in the corresponding layer or likelihood nodes in the final layer) matrices named output1, output2,..., outputD. Each matrix gives samples of the output from one of D GPs/likelihoods at the testing positions, and has its rows corresponding to testing positions and columns corresponding to samples of size: N * sample_size, where N is the number of imputations specified in emulator().

(Linked (D)GP case) If obj is produced by lgp():

if method = "mean_var" and full_layer = FALSE: a named list is returned. The list contains two sub-lists named mean for the predictive means and var for the predictive variances respectively. Each sub-list contains M number (same number of computer models in the final layer of the system) of matrices named emulator1, emulator2,..., emulatorM. Each matrix has its rows corresponding to global testing positions and columns corresponding to output dimensions of the associated emulator in the final layer.

if method = "mean_var" and full_layer = TRUE: a named list is returned. The list contains two sub-lists named mean for the predictive means and var for the predictive variances respectively. Each sub-list contains L (i.e., the number of layers of the emulated system) components named layer1, layer2,..., layerL. Each component represents a layer and contains M number (same number of emulators in the corresponding layer of the system) of matrices named emulator1, emulator2,..., emulatorM. Each matrix has its rows corresponding to global testing positions and columns corresponding to output dimensions of the associated GP/DGP emulator in the corresponding layer.

if method = "sampling" and full_layer = FALSE: a named list is returned. The list contains M number (same number of emulators in the final layer of the system) of sub-lists named emulator1, emulator2,..., emulatorM. Each sub-list corresponds to an emulator in the final layer, and contains D matrices, named output1, output2,..., outputD, that correspond to the output dimensions of the GP/DGP emulator. Each matrix has its rows corresponding to testing positions and columns corresponding to samples of size: N * sample_size, where N is the number of imputations specified in emulator().

if method = "sampling" and full_layer = TRUE: a named list is returned. The list contains L (i.e., the number of layers of the emulated system) sub-lists named layer1, layer2,..., layerL. Each sub-list represents a layer and contains M number (same number of emulators in the corresponding layer of the system) of components named emulator1, emulator2,..., emulatorM. Each component corresponds to an emulator in the associated layer, and contains D matrices, named output1, output2,..., outputD, that correspond to the output dimensions of the GP/DGP emulator. Each matrix has its rows corresponding to testing positions and columns corresponding to samples of size: N * sample_size, where N is the number of imputations specified in emulator().